requirements.txt
Django==5.1.15
djangorestframework==3.17.1
python-decouple==3.8
PyMySQL==1.1.2
django-session-timeout==0.1.0
django-crispy-forms==1.14.0
django-debug-toolbar==2.2
django-multiselectfield==1.0.1
django-autocomplete-light==3.12.1
django-betterforms==3.0.0
django-filter==25.2
django-select2==8.4.8
django-appconf==1.2.0
Pillow==12.2.0
openpyxl==3.1.5
reportlab==4.4.10
xhtml2pdf==0.2.17
PyPDF2==3.0.1
pandas==3.0.2
numpy==2.4.4
xlrd==2.0.2
legacy-cgi==2.6.4
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
pytz==2026.1.post1
requests==2.33.1
